backend/requirements.txt
asgiref==3.6.0
asyncio==3.4.3
certifi==2022.12.7
charset-normalizer==3.1.0
click==8.1.3
Django==4.2
django-cors-headers==3.14.0
djangorestframework==3.14.0
idna==3.4
joblib==1.2.0
nltk==3.8.1
pytz==2023.3
ranger-fm==1.9.3
regex==2023.3.23
requests==2.28.2
sqlparse==0.4.3
team==1.0
tqdm==4.65.0
urllib3==1.26.15
